refactor: rename nodes to be redis1, redis2, redis3

+++ provision/redis/3-redis-commander.sh
@@ -1,17 +1,20 @@
 #!/bin/bash
 
-# use full for redis-commander
-yum install nodejs npm -y
-npm install -g redis-commander
+FLAG_FILE="/etc/vagrant_provisioned"
 
-# Add more paths to the PATH variable
-if ! grep -q "export PATH=\$PATH:/usr/local/bin" ~/.bashrc; then
-    # If not present, append the line to .bashrc
-    echo 'export PATH=$PATH:/usr/local/bin' >> ~/.bashrc
+# Only once...
+if [ ! -f "$FLAG_FILE" ]; then
+
+  curl -sL https://rpm.nodesource.com/setup_18.x | bash -
+  yum install nodejs -y
+  npm install -g npm@latest
+  npm install -g redis-commander
+
+  cp /tmp/local-development.json /usr/lib/node_modules/redis-commander/config/local-development.json
+  cp /tmp/redis-commander.service /etc/systemd/system/redis-commander.service
+  systemctl daemon-reload
+  systemctl enable redis-commander.service -q
+  systemctl restart redis-commander.service
+
+  touch "$FLAG_FILE"
 fi
-
-cp /tmp/local-development.json /usr/local/lib/node_modules/redis-commander/config/local-development.json
-cp /tmp/redis-commander.service /etc/systemd/system/redis-commander.service
-systemctl daemon-reload
-systemctl enable redis-commander.service -q
-systemctl restart redis-commander.service
